requirements.txt
duckduckgo-search
duckduckgo-images-api
fastai
fastbook
graphviz
jupyterlab
keras
matplotlib
mypy
numpy
nose2
pandas
Pillow
pip
seaborn
torch
torchvision
